requirements.txt
discord.py==2.4.0
pytz==2024.2
emoji==2.14.1
GitPython==3.1.43
